requirements.txt
asgiref==3.5.0
Django==4.0.3
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
mysqlclient==2.1.0
Pillow==9.1.0
sqlparse==0.4.2
tzdata==2022.1
